# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Graph Platform API
  version: 3.2.0
  description: API Graph Platform with Graph, FSM, Review, Ontology, RAG, Execution and Copilot engines.
servers:
  - url: http://localhost:3001/api
security:
  - bearerAuth: []
paths:
...
  /copilot/chat:
    post: { operationId: copilotChat, security: [], requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '200': { $ref: '#/components/responses/Success' } } }
...
  /conversations:
    get: { operationId: listConversations, responses: { '200': { $ref: '#/components/responses/Success' } } }
    post: { operationId: createConversation, requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '201': { $ref: '#/components/responses/Success' } } }
  /conversations/{id}:
    get: { operationId: getConversation, parameters: [{ $ref: '#/components/parameters/Id' }], responses: { '200': { $ref: '#/components/responses/Success' } } }
  /conversations/{id}/decisions:
    get: { operationId: listDecisions, parameters: [{ $ref: '#/components/parameters/Id' }], responses: { '200': { $ref: '#/components/responses/Success' } } }
    post: { operationId: createDecision, parameters: [{ $ref: '#/components/parameters/Id' }], requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '201': { $ref: '#/components/responses/Success' } } }
...
  /answers/{id}/feedback:
    patch: { operationId: answerFeedback, parameters: [{ $ref: '#/components/parameters/Id' }], requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '200': { $ref: '#/components/responses/Success' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  responses:
    Success:
      description: Successful response
      content:
        application/json:
          schema: { type: [object, array], additionalProperties: true }